Add new clustering prediction script and remove old root reco script
Introduced `save_pred_mlclustering.py`, a new script for saving predictions related to machine learning clustering tasks. This script includes functions for energy prediction calculations and hit-query assignments, enhancing the clustering workflow. Additionally, removed the outdated `save_root_reco.py` script to streamline the codebase and eliminate redundancy.
