+++ backups/modifications/objectcondensation_.py
